+++ mapjs/src/start.js
@@ -71,6 +71,15 @@ const layoutThemeStyle = function (themeJson) {
 	return true;
 };
 
+// TODO: add this somewhere else. Should be called by listener and other functions like initialisation and drag and drop?
+// 	Or could add it to mapModel class and then don't need the container argument at all
+window.loadMap = function (mapJson) {
+	// For some reason, first argument in call: container being passed as this. Not sure why
+	// TODO: Need to use correct mapModel
+	//	 mapInstance[this.target_container_id].mapModel.getIdea();
+	this.mapModel.setIdea(content(JSON.parse(mapJson)));
+}
+
 // Changes theme of all maps on page
 // 	To change, call e.g. changeTheme(map, MAPJS.v1)
 // 	TODO: Move to mapModel.changeTheme() in map-model.js
@@ -198,6 +207,34 @@ const addMap = function (container, mapJson) {
 			}
 		}
 	});
+
+	// Reference: Old drag and drop
+	// container.on('drop', function (e) {
+	// 	const dataTransfer = e.originalEvent.dataTransfer;
+	// 	e.stopPropagation();
+	// 	e.preventDefault();
+	// 	if (dataTransfer && dataTransfer.files && dataTransfer.files.length > 0) {
+	// 		const fileInfo = dataTransfer.files[0];
+	// 		if (/\.mup$/.test(fileInfo.name)) {
+	// 			const oFReader = new window.FileReader();
+	// 			oFReader.onload = function (oFREvent) {
+	// 				mapModel.setIdea(content(JSON.parse(oFREvent.target.result)));
+	// 			};
+	// 			oFReader.readAsText(fileInfo, 'UTF-8');
+	// 		}
+	// 	}
+	// });
+
 };
 
 document.addEventListener('DOMContentLoaded', init);
+
+// Hacky way to test
+window.setTimeout(function () {
+	b = document.getElementById("submit");
+	if (b) {
+		// Automatically click button to speed up testing
+		b.click();
+	}
+}, 400 /* but after 400 ms */);
+